# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Dokumentasi API Terpadu SIM Sekolah SD Negeri.",
        "title": "SIM Sekolah API",
        "termsOfService": "http://example.com/terms/",
        "contact": {
            "name": "Developer Team",
            "email": "admin@simsekolah.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/academic/graduation": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Promotion Management"
                ],
                "summary": "Graduate Students Massively",
                "parameters": [
                    {
                        "description": "Graduation Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_promotion.GraduateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/sim-sekolah_internal_common.Response"
                        }
                    }
                }
            }
        },
        "/academic/promotion": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Promotion Management"
                ],
                "summary": "Promote Students to Next Classroom",
                "parameters": [
                    {
                        "description": "Promotion Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_promotion.PromoteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/sim-sekolah_internal_common.Response"
                        }
                    }
                }
            }
        },
...
        "/teachers": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Ambil semua data guru dengan pencarian",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Teacher"
                ],
                "summary": "Get All Teachers",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search by name, NIP, email",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/sim-sekolah_internal_common.PaginationResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Teacher"
                ],
                "summary": "Create Teacher",
                "parameters": [
                    {
                        "description": "Create Teacher",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_teacher.CreateTeacherRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/sim-sekolah_internal_common.Response"
                        }
                    }
                }
            }
        },
        "/teachers/school/{schoolId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Teacher"
                ],
                "summary": "Get Teachers By School",
                "parameters": [
                    {
                        "type": "string",
                        "description": "School ID",
                        "name": "schoolId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/sim-sekolah_internal_common.Response"
                        }
                    }
                }
            }
        },
        "/teachers/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Teacher"
                ],
                "summary": "Get Teacher By ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Teacher ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/sim-sekolah_internal_common.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Teacher"
                ],
                "summary": "Update Teacher",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Teacher ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Teacher",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_teacher.UpdateTeacherRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/sim-sekolah_internal_common.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Teacher"
                ],
                "summary": "Delete Teacher",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Teacher ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/sim-sekolah_internal_common.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_promotion.GraduateRequest": {
            "type": "object",
            "required": [
                "student_ids"
            ],
            "properties": {
                "student_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "internal_promotion.PromoteRequest": {
            "type": "object",
            "required": [
                "source_classroom_id",
                "student_ids",
                "target_classroom_id"
            ],
            "properties": {
                "source_classroom_id": {
                    "type": "string"
                },
                "student_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "target_classroom_id": {
                    "type": "string"
                }
            }
        },
...
        "internal_teacher.CreateTeacherRequest": {
            "type": "object",
            "required": [
                "full_name",
                "school_id"
            ],
            "properties": {
                "additional_position": {
                    "description": "Tugas tambahan di luar mengajar (e.g. Kepala Laboratorium)",
                    "type": "string",
                    "example": "Kepala Perpustakaan"
                },
                "appointment_decree": {
                    "description": "Nomor SK Pengangkatan pertama",
                    "type": "string",
                    "example": "800/234/SK-KADIS/2010"
                },
                "birth_date": {
                    "description": "Tanggal lahir format YYYY-MM-DD",
                    "type": "string",
                    "example": "1985-05-12"
                },
                "birth_place": {
                    "description": "Tempat lahir guru",
                    "type": "string",
                    "example": "Makassar"
                },
                "certificate_number": {
                    "description": "Nomor sertifikasi pendidik jika tersertifikasi",
                    "type": "string",
                    "example": "10234/CERT/2012"
                },
                "district": {
                    "description": "Nama kecamatan",
                    "type": "string",
                    "example": "Pasimarannu"
                },
                "email": {
                    "description": "Alamat email resmi aktif",
                    "type": "string",
                    "maxLength": 100,
                    "example": "budi@sd.sch.id"
                },
                "employment_status": {
                    "description": "Status kepegawaian aktif",
                    "type": "string",
                    "enum": [
                        "PNS",
                        "PPPK",
                        "Honorer",
                        "GTY"
                    ],
                    "example": "PNS"
                },
                "full_address": {
                    "description": "Kontak \u0026 Alamat",
                    "type": "string",
                    "example": "Jl. Trans Sulawesi No. 85"
                },
                "full_name": {
                    "description": "Identitas",
                    "type": "string",
                    "maxLength": 255,
                    "example": "Budi Santoso, S.Pd"
                },
                "gender": {
                    "description": "Jenis kelamin: 'L' (Laki-laki) atau 'P' (Perempuan)",
                    "type": "string",
                    "enum": [
                        "L",
                        "P"
                    ],
                    "example": "L"
                },
                "graduation_year": {
                    "description": "Tahun kelulusan kuliah",
                    "type": "integer",
                    "example": 2008
                },
                "hamlet": {
                    "description": "Nama dusun / lingkungan",
                    "type": "string",
                    "example": "Dusun Bonerate"
                },
                "is_active": {
                    "description": "Status keaktifan guru di sekolah",
                    "type": "boolean",
                    "example": true
                },
                "is_certified": {
                    "description": "Status sertifikasi pendidik",
                    "type": "boolean",
                    "example": true
                },
                "last_education": {
                    "description": "Pendidikan",
                    "type": "string",
                    "example": "S1"
                },
                "major": {
                    "description": "Jurusan pendidikan terakhir",
                    "type": "string",
                    "example": "Pendidikan Matematika"
                },
                "nationality": {
                    "description": "Kewarganegaraan (e.g. WNI, WNA)",
                    "type": "string",
                    "example": "WNI"
                },
                "nik": {
                    "description": "Nomor Induk Kependudukan (16 digit)",
                    "type": "string",
                    "maxLength": 20,
                    "example": "7301021205850001"
                },
                "nip": {
                    "description": "Kepegawaian",
                    "type": "string",
                    "maxLength": 30,
                    "example": "198505122010011003"
                },
                "niy_nigk": {
                    "description": "Nomor Induk Yayasan / GTY jika sekolah swasta",
                    "type": "string",
                    "example": "NIY-098765"
                },
                "nuptk": {
                    "description": "Nomor Unik Pendidik dan Tenaga Kependidikan (16 digit)",
                    "type": "string",
                    "maxLength": 20,
                    "example": "9876543210123456"
                },
                "phone": {
                    "description": "Nomor telepon seluler aktif",
                    "type": "string",
                    "maxLength": 30,
                    "example": "081234567890"
                },
                "photo_url": {
                    "description": "URL berkas foto profil guru",
                    "type": "string",
                    "example": "https://example.com/photos/budi.jpg"
                },
                "postal_code": {
                    "description": "Kode pos alamat domisili",
                    "type": "string",
                    "example": "92854"
                },
                "province": {
                    "description": "Nama provinsi",
                    "type": "string",
                    "example": "Sulawesi Selatan"
                },
                "regency": {
                    "description": "Nama kabupaten atau kota",
                    "type": "string",
                    "example": "Kepulauan Selayar"
                },
                "religion": {
                    "description": "Agama guru",
                    "type": "string",
                    "example": "Islam"
                },
                "rt_rw": {
                    "description": "Rukun Tetangga / Rukun Warga",
                    "type": "string",
                    "example": "002/001"
                },
                "salary_source": {
                    "description": "Sumber gaji utama (APBN, APBD, Yayasan)",
                    "type": "string",
                    "example": "APBD Kabupaten"
                },
                "school_id": {
                    "description": "UUID Sekolah tempat guru mengajar",
                    "type": "string",
                    "example": "804b75d4-2822-48a7-8eba-3566f038f81f"
                },
                "start_teaching_date": {
                    "description": "Tanggal mulai mengajar pertama kali format YYYY-MM-DD",
                    "type": "string",
                    "example": "2010-01-02"
                },
                "teaching_hours": {
                    "description": "Total jam mengajar tatap muka per minggu",
                    "type": "integer",
                    "example": 24
                },
                "teaching_preference": {
                    "description": "Narasi AI",
                    "type": "string",
                    "example": "Sangat sabar dan menyukai pembelajaran berbasis gamifikasi"
                },
                "teaching_subject": {
                    "description": "Mata pelajaran utama yang diampu",
                    "type": "string",
                    "example": "Matematika"
                },
                "university_name": {
                    "description": "Nama perguruan tinggi kelulusan",
                    "type": "string",
                    "example": "Universitas Negeri Makassar"
                },
                "village": {
                    "description": "Nama desa atau kelurahan",
                    "type": "string",
                    "example": "Bonerate"
                }
            }
        },
        "internal_teacher.UpdateTeacherRequest": {
            "type": "object",
            "properties": {
                "additional_position": {
                    "type": "string",
                    "example": "Kepala Perpustakaan"
                },
                "appointment_decree": {
                    "type": "string",
                    "example": "800/234/SK-KADIS/2010"
                },
                "birth_date": {
                    "description": "Tanggal lahir YYYY-MM-DD",
                    "type": "string",
                    "example": "1985-05-12"
                },
                "birth_place": {
                    "description": "Tempat lahir",
                    "type": "string",
                    "example": "Makassar"
                },
                "certificate_number": {
                    "type": "string",
                    "example": "10234/CERT/2012"
                },
                "district": {
                    "type": "string",
                    "example": "Pasimarannu"
                },
                "email": {
                    "type": "string",
                    "example": "budi@sd.sch.id"
                },
                "employment_status": {
                    "type": "string",
                    "enum": [
                        "PNS",
                        "PPPK",
                        "Honorer",
                        "GTY"
                    ],
                    "example": "PNS"
                },
                "full_address": {
                    "description": "Kontak \u0026 Alamat",
                    "type": "string",
                    "example": "Jl. Trans Sulawesi No. 85"
                },
                "full_name": {
                    "description": "Identitas",
                    "type": "string",
                    "example": "Budi Santoso, S.Pd, M.Pd"
                },
                "gender": {
                    "description": "Jenis kelamin",
                    "type": "string",
                    "enum": [
                        "L",
                        "P"
                    ],
                    "example": "L"
                },
                "graduation_year": {
                    "type": "integer",
                    "example": 2008
                },
                "hamlet": {
                    "type": "string",
                    "example": "Dusun Bonerate"
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "is_certified": {
                    "type": "boolean",
                    "example": true
                },
                "last_education": {
                    "description": "Pendidikan",
                    "type": "string",
                    "example": "S1"
                },
                "major": {
                    "type": "string",
                    "example": "Pendidikan Matematika"
                },
                "nationality": {
                    "description": "Kewarganegaraan",
                    "type": "string",
                    "example": "WNI"
                },
                "nik": {
                    "description": "Nomor Induk Kependudukan",
                    "type": "string",
                    "example": "7301021205850001"
                },
                "nip": {
                    "description": "Kepegawaian",
                    "type": "string",
                    "example": "198505122010011003"
                },
                "niy_nigk": {
                    "description": "Nomor NIY/NIGK",
                    "type": "string",
                    "example": "NIY-098765"
                },
                "nuptk": {
                    "description": "Nomor NUPTK (16 digit)",
                    "type": "string",
                    "example": "9876543210123456"
                },
                "phone": {
                    "type": "string",
                    "example": "081234567890"
                },
                "photo_url": {
                    "description": "Foto profil URL",
                    "type": "string",
                    "example": "https://example.com/photos/budi.jpg"
                },
                "postal_code": {
                    "type": "string",
                    "example": "92854"
                },
                "province": {
                    "type": "string",
                    "example": "Sulawesi Selatan"
                },
                "regency": {
                    "type": "string",
                    "example": "Kepulauan Selayar"
                },
                "religion": {
                    "description": "Agama",
                    "type": "string",
                    "example": "Islam"
                },
                "rt_rw": {
                    "type": "string",
                    "example": "002/001"
                },
                "salary_source": {
                    "type": "string",
                    "example": "APBD Kabupaten"
                },
                "start_teaching_date": {
                    "type": "string",
                    "example": "2010-01-02"
                },
                "teaching_hours": {
                    "type": "integer",
                    "example": 24
                },
                "teaching_preference": {
                    "description": "Narasi AI",
                    "type": "string",
                    "example": "Sangat sabar dan menyukai pembelajaran berbasis gamifikasi"
                },
                "teaching_subject": {
                    "type": "string",
                    "example": "Matematika"
                },
                "university_name": {
                    "type": "string",
                    "example": "Universitas Negeri Makassar"
                },
                "village": {
                    "type": "string",
                    "example": "Bonerate"
                }
            }
        },
...
        "sim-sekolah_internal_common.PaginationResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "limit": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "page": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "total_pages": {
                    "type": "integer"
                },
                "total_rows": {
                    "type": "integer"
                }
            }
        },
        "sim-sekolah_internal_common.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}